requirements.txt
# requirements.txt
streamlit>=1.31.0
supabase>=2.3.0
together>=0.2.0
python-dotenv>=1.0.0
